Move type helpers to core/types folder
