+++ include/ct/math/detail/tvec2.hpp
@@ -4,7 +4,7 @@
 
 #include <ostream>
 #include <ct/math/detail/setup.hpp>
-#include <ct/core/utils/type_info.hpp>
+#include <ct/core/types/type_info.hpp>
 
 namespace ct {
 
@@ -91,9 +91,7 @@ namespace ct {
 
 		CT_CONSTEXPR T & operator[](const u32 & index);
 		CT_CONSTEXPR const T & operator[](const u32 & index) const;
-
 	};
-
 
 	// Unary operators
 
@@ -166,7 +164,7 @@ namespace ct {
 	// ostream operator
 
 	template <typename T>
-	std::ostream& operator<<(std::ostream &s, const tvec2<T> &v);
+	std::ostream & operator<<(std::ostream & s, const tvec2<T> & v);
 
 } // namespace ct
