+++ include/ct/math/detail/tvec3.hpp
@@ -3,7 +3,7 @@
 #define _CHRONOTRIX_FRWK_MATH_TVEC3_HPP_
 
 #include <ct/math/detail/setup.hpp>
-#include <ct/core/utils/type_info.hpp>
+#include <ct/core/types/type_info.hpp>
 
 namespace ct {
 
@@ -121,7 +121,6 @@ namespace ct {
 
 		CT_CONSTEXPR T & operator[](const u32 & index);
 		CT_CONSTEXPR const T & operator[](const u32 & index) const;
-
 	};
 
 	// Unary operators
@@ -195,9 +194,9 @@ namespace ct {
 	// ostream operator
 
 	template <typename T>
-	std::ostream& operator<<(std::ostream &s, const tvec3<T> &v);
+	std::ostream & operator<<(std::ostream & s, const tvec3<T> & v);
 
-}
+} // namespace ct
 
 #include <ct/math/detail/tvec3.inl>
